Refactor/508 change variable in StorageUnit from long int to int64_t

+++ src/simulation/onboardDataHandling/storageUnit/simpleStorageUnit.h
@@ -30,7 +30,7 @@ class SimpleStorageUnit: public DataStorageUnitBase {
 public:
     SimpleStorageUnit();
     ~SimpleStorageUnit();
-    void setDataBuffer(long long int data); //!< Method to add/remove data from the storage unit once
+    void setDataBuffer(int64_t data); //!< Method to add/remove data from the storage unit once
 
 private:
     void customReset(uint64_t CurrentClock); //!< Custom Reset method
